fixing the delete comment, preventing the delete twice
git-svn-id: file:///home/jaliste/hcvol/hcvol/trunk@3085 4f49b16c-3c57-40ef-8e22-7452baa6c33a

+++ htdocs/bw/admin/admincomments.php
@@ -125,10 +125,16 @@ switch ($action) {
 		}
 		$Message = " Delete comment #" . GetParam("IdComment");
 		$c = LoadRow("select * from comments where id=" . GetParam("IdComment"));
+		if (!isset($c->id)) {
+		   $Message=" No such coment" ;
+		   DisplayAdminComments(loaddata("", " and comments.id=" . GetParam("IdComment")), $Message); // call the layout
+		   exit (0);
+		   break ;
+		}
 		$str = "delete from comments  where id=" . GetParam("IdComment");
 		sql_query($str);
 		LogStr("Deleting comment #" . GetParam("IdComment") . " previous where=" . $c->TextWhere . " previous text=" . $c->TextFree . " previous Quality=" . $c->Quality, "AdminComment");
-		DisplayAdminComments(loaddata("", " and comments.IdMember=".$c->IdToMember ), $Message); // call the layout
+		DisplayAdminComments(loaddata("", " and comments.IdToMember=".$c->IdToMember ), $Message); // call the layout
 		exit (0);
 		break;
